renamed validation_traversal.rs to conditions_validation.rs
- It's better separate from the actual validation traversal module.
